forms
ugly checkboxes now used for profile team selection. also able to save
this info to profile form.
this can be applied to friends and bars in different ways. still would
be nice to put profile creation on another page. an "edit profile page"
that could be revisited later by the user.