setting up game page
game page mirrors bar page functions essentially.
show events around this game, show attendees for those events, tell user
if they are registered for an event for this game, give option to create
an event around this game or join an existing event.

need to add "join event" to bar page too. this logic can be reused for
invitations.