got a team page up
rolling right along.
need to create a "watch game" page to create an event coming from the
game. user will have to select a bar.
ideally this will be a search bar with a populated list of results below
it. top of list initially will be user's favorite bars. as the enter
search (onkeyup) suggestions will be made matching the query from the
DB. this logic will open doors to functionality elsewhere and will be
important to overall usability.